More api surface type cleanup
