+++ src/actions/actions.types.ts
@@ -1,60 +1,7 @@
 import type { ModifierStage } from "../processTargets/PipelineStages.types";
 import type { Target } from "../typings/target.types";
 import type { SelectionWithEditor } from "../typings/Types";
-
-export type ActionType =
-  | "callAsFunction"
-  | "clearAndSetSelection"
-  | "copyToClipboard"
-  | "cutToClipboard"
-  | "deselect"
-  | "editNew"
-  | "editNewLineAfter"
-  | "editNewLineBefore"
-  | "executeCommand"
-  | "extractVariable"
-  | "findInWorkspace"
-  | "foldRegion"
-  | "followLink"
-  | "generateSnippet"
-  | "getText"
-  | "highlight"
-  | "indentLine"
-  | "insertCopyAfter"
-  | "insertCopyBefore"
-  | "insertEmptyLineAfter"
-  | "insertEmptyLineBefore"
-  | "insertEmptyLinesAround"
-  | "insertSnippet"
-  | "moveToTarget"
-  | "outdentLine"
-  | "pasteFromClipboard"
-  | "randomizeTargets"
-  | "remove"
-  | "rename"
-  | "replace"
-  | "replaceWithTarget"
-  | "revealDefinition"
-  | "revealTypeDefinition"
-  | "reverseTargets"
-  | "rewrapWithPairedDelimiter"
-  | "scrollToBottom"
-  | "scrollToCenter"
-  | "scrollToTop"
-  | "setSelection"
-  | "setSelectionAfter"
-  | "setSelectionBefore"
-  | "showDebugHover"
-  | "showHover"
-  | "showQuickFix"
-  | "showReferences"
-  | "sortTargets"
-  | "swapTargets"
-  | "toggleLineBreakpoint"
-  | "toggleLineComment"
-  | "unfoldRegion"
-  | "wrapWithPairedDelimiter"
-  | "wrapWithSnippet";
+import type { ActionType } from "../core/commandRunner/typings/ActionCommand";
 
 /**
  * To be returned by {@link Action.run}
