+++ src/core/commandRunner/typings/ActionCommand.ts
@@ -1,4 +1,56 @@
-import { ActionType } from "../../../actions/actions.types";
+export type ActionType =
+  | "callAsFunction"
+  | "clearAndSetSelection"
+  | "copyToClipboard"
+  | "cutToClipboard"
+  | "deselect"
+  | "editNew"
+  | "editNewLineAfter"
+  | "editNewLineBefore"
+  | "executeCommand"
+  | "extractVariable"
+  | "findInWorkspace"
+  | "foldRegion"
+  | "followLink"
+  | "generateSnippet"
+  | "getText"
+  | "highlight"
+  | "indentLine"
+  | "insertCopyAfter"
+  | "insertCopyBefore"
+  | "insertEmptyLineAfter"
+  | "insertEmptyLineBefore"
+  | "insertEmptyLinesAround"
+  | "insertSnippet"
+  | "moveToTarget"
+  | "outdentLine"
+  | "pasteFromClipboard"
+  | "randomizeTargets"
+  | "remove"
+  | "rename"
+  | "replace"
+  | "replaceWithTarget"
+  | "revealDefinition"
+  | "revealTypeDefinition"
+  | "reverseTargets"
+  | "rewrapWithPairedDelimiter"
+  | "scrollToBottom"
+  | "scrollToCenter"
+  | "scrollToTop"
+  | "setSelection"
+  | "setSelectionAfter"
+  | "setSelectionBefore"
+  | "showDebugHover"
+  | "showHover"
+  | "showQuickFix"
+  | "showReferences"
+  | "sortTargets"
+  | "swapTargets"
+  | "toggleLineBreakpoint"
+  | "toggleLineComment"
+  | "unfoldRegion"
+  | "wrapWithPairedDelimiter"
+  | "wrapWithSnippet";
 
 export interface ActionCommand {
   /**
