+++ src/core/commandRunner/typings/legacy/CommandV2.types.ts
@@ -1,5 +1,58 @@
-import type { ActionType } from "../../../../actions/actions.types";
 import type { PartialTargetDescriptorV2 } from "./targetDescriptorV2.types";
+
+type ActionType =
+  | "callAsFunction"
+  | "clearAndSetSelection"
+  | "copyToClipboard"
+  | "cutToClipboard"
+  | "deselect"
+  | "editNew"
+  | "editNewLineAfter"
+  | "editNewLineBefore"
+  | "executeCommand"
+  | "extractVariable"
+  | "findInWorkspace"
+  | "foldRegion"
+  | "followLink"
+  | "generateSnippet"
+  | "getText"
+  | "highlight"
+  | "indentLine"
+  | "insertCopyAfter"
+  | "insertCopyBefore"
+  | "insertEmptyLineAfter"
+  | "insertEmptyLineBefore"
+  | "insertEmptyLinesAround"
+  | "insertSnippet"
+  | "moveToTarget"
+  | "outdentLine"
+  | "pasteFromClipboard"
+  | "randomizeTargets"
+  | "remove"
+  | "rename"
+  | "replace"
+  | "replaceWithTarget"
+  | "revealDefinition"
+  | "revealTypeDefinition"
+  | "reverseTargets"
+  | "rewrapWithPairedDelimiter"
+  | "scrollToBottom"
+  | "scrollToCenter"
+  | "scrollToTop"
+  | "setSelection"
+  | "setSelectionAfter"
+  | "setSelectionBefore"
+  | "showDebugHover"
+  | "showHover"
+  | "showQuickFix"
+  | "showReferences"
+  | "sortTargets"
+  | "swapTargets"
+  | "toggleLineBreakpoint"
+  | "toggleLineComment"
+  | "unfoldRegion"
+  | "wrapWithPairedDelimiter"
+  | "wrapWithSnippet";
 
 interface ActionCommand {
   /**
