+++ src/core/commandRunner/typings/legacy/targetDescriptorV2.types.ts
@@ -1,4 +1,30 @@
-import { HatStyleName } from "../../../hatStyles";
+const HAT_COLORS = [
+  "default",
+  "blue",
+  "green",
+  "red",
+  "pink",
+  "yellow",
+  "userColor1",
+  "userColor2",
+] as const;
+
+const HAT_NON_DEFAULT_SHAPES = [
+  "ex",
+  "fox",
+  "wing",
+  "hole",
+  "frame",
+  "curve",
+  "eye",
+  "play",
+  "bolt",
+  "crosshairs",
+] as const;
+
+type HatColor = typeof HAT_COLORS[number];
+type HatNonDefaultShape = typeof HAT_NON_DEFAULT_SHAPES[number];
+type HatStyleName = HatColor | `${HatColor}-${HatNonDefaultShape}`;
 
 interface CursorMark {
   type: "cursor";
