+++ src/core/commandVersionUpgrades/canonicalizeTargets.ts
@@ -6,7 +6,7 @@ import {
   SimpleScopeTypeType,
 } from "../commandRunner/typings/targetDescriptor.types";
 import { transformPartialPrimitiveTargets } from "../../util/getPrimitiveTargets";
-import { HatStyleName } from "../hatStyles";
+import { HatStyleName } from "../commandRunner/typings/hatStyles.types";
 
 const SCOPE_TYPE_CANONICALIZATION_MAPPING: Record<string, SimpleScopeTypeType> =
   {
